+++ framer.h
@@ -5,6 +5,6 @@
 #include "miniv.h"
 #pragma once
 
-err_t frameReadLen(int fd, int *len);
-err_t frameWriteLen(int fd, int len);
+err_t frameReadLen(int fd, unsigned long *len);
+err_t frameWriteLen(int fd, unsigned long len);
 err_t frameWrite(int fd, buf_t *b);
